+++ packages/core/src/components/TextField/Styled/HelperAndErrorTextTooltipWrapper.styled.tsx
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
 
-export const HelperAndErrorTextTooltipWrapper = styled.div<{ displayCharacterCount: boolean; size: 'S' | 'M' }>`
-    ${({ displayCharacterCount, size }) => (!displayCharacterCount || size !== 'S') && 'margin-left: 1rem;'}
+export const HelperAndErrorTextTooltipWrapper = styled.div<{ displayCharacterCount: boolean; size: 'S' | 'M'; isSuffixPresent: boolean; }>`
+    margin-top: 0.5rem;
+    margin-right: ${({ isSuffixPresent }) => !isSuffixPresent && '1rem'};
+    margin-left: ${({ displayCharacterCount, size }) => (!displayCharacterCount || size !== 'S' ) && '1rem'};
 `;