+++ packages/utils/src/helpers/cookieStorage/index.ts
@@ -1,1 +1,1 @@
-export { COOKIE_STORAGE as default } from './cookieStorage';
+export * from './cookieStorage';