+++ packages/utils/src/hooks/useStorage/useStorage.ts
@@ -1,7 +1,7 @@
-import { useContext, useEffect, useState } from 'react';
-import COOKIE_STORAGE from '../../helpers/cookieStorage';
-import LOCAL_STORAGE from '../../helpers/localStorage';
-import SESSION_STORAGE from '../../helpers/sessionStorage';
+import { useContext, useEffect, useRef, useState } from 'react';
+import { COOKIE_STORAGE } from '../../helpers/cookieStorage';
+import { LOCAL_STORAGE } from '../../helpers/localStorage';
+import { SESSION_STORAGE } from '../../helpers/sessionStorage';
 import { StorageConfig } from './Storage.context';
 import { UseStorageOptions, UseStorageSetValue } from './types';
 
@@ -15,7 +15,8 @@ export const useStorage = <T>(key: string, currOptions?: UseStorageOptions<T>): 
     const contextOption: UseStorageOptions<T> = useContext(StorageConfig),
         options = currOptions ?? contextOption,
         storage = storageObj[options.storage || 'localStorage'],
-        initialValue = options.initialValue;
+        initialValue = options.initialValue,
+        isMounted = useRef<boolean>(true);
 
     const readValue = (): T => {
         if (typeof window === 'undefined') {
@@ -24,7 +25,7 @@ export const useStorage = <T>(key: string, currOptions?: UseStorageOptions<T>): 
         }
 
         const item = storage.getItem(key);
-        if (!item) {
+        if (item === null && initialValue !== undefined) {
             storage.setItem(key, initialValue);
             window?.dispatchEvent(new Event(`local-storage-${key}`));
             return initialValue;
@@ -41,15 +42,22 @@ export const useStorage = <T>(key: string, currOptions?: UseStorageOptions<T>): 
 
         try {
             storage.setItem(key, value);
-            setStoredValue(value);
+            isMounted.current && setStoredValue(value);
             window.dispatchEvent(new Event(`local-storage-${key}`));
         } catch (error) {
             console.warn(`Error setting localStorage key “${key}”:`, error);
         }
     };
 
+    useEffect(
+        () => () => {
+            isMounted.current = false;
+        },
+        []
+    );
+
     useEffect(() => {
-        const handleStorageChange = () => setStoredValue(readValue());
+        const handleStorageChange = () => isMounted.current && setStoredValue(readValue());
 
         window?.addEventListener(`local-storage-${key}`, handleStorageChange);
         return () => window.removeEventListener('local-storage', handleStorageChange);